refactor(storage)!: P6.2 land canonical InMemoryStore impl in nebula-storage
Populate crates/storage/src/credential/memory.rs with the canonical
`InMemoryStore` impl (ADR-0029 §2 amended by ADR-0032 — trait stays in
credential; impls move to storage). Gated behind `credential-in-memory`
feature; re-exported from `nebula_storage::credential::InMemoryStore`.

Credential keeps a behaviour-identical copy of `InMemoryStore` in
`crates/credential/src/store_memory.rs` because its own internal
consumers (resolver tests, layer tests, integration tests under
`crates/credential/tests/`) reference it directly. They cannot reach
`nebula_storage::credential::InMemoryStore` — ADR-0032 §3 forbids
`nebula-credential → nebula-storage` in either `[dependencies]` or
`[dev-dependencies]` (the dev-dep form triggers a "two copies of
crate" cargo resolution that breaks the `CredentialStore` trait bound).

`test_helpers::make_credential` in `crates/credential/src/store.rs`
is now `pub` under `#[cfg(any(test, feature = "test-util"))]` so
storage's `memory.rs` tests can import it. `nebula-storage`'s
`test-util` feature forwards to `nebula-credential/test-util` to
propagate the gate, and the test module in `storage/src/credential/
memory.rs` is gated `#[cfg(all(test, feature = "test-util"))]` so
default-feature clippy --all-targets stays clean.

Dep-graph stays strictly one-directional at both prod and dev level:
  nebula-storage -> nebula-credential
Verified via `cargo tree -p nebula-credential` (no storage edge).

Tests:
- nebula-storage --features credential-in-memory,test-util: 84/84 PASS
- nebula-credential --all-features: 428/432 PASS (4 pre-existing
  rotation::events::tests::* failures carry over from P1 — same
  `WrongPrefix { expected_prefix: "cred" }` errors reproduce on the
  parent commit 3ea4c505; tracked for P8).

Ref: docs/adr/0032-credential-store-canonical-home.md
