+++ crates/storage/src/credential/mod.rs
@@ -13,11 +13,11 @@ pub mod store;
 #[cfg(any(test, feature = "credential-in-memory"))]
 pub mod memory;
 
-// TODO(P6.2-P6.5): re-export once files are populated.
+#[cfg(any(test, feature = "credential-in-memory"))]
+pub use memory::InMemoryStore;
+
+// TODO(P6.3-P6.5): re-export once files are populated.
 // pub use key_provider::{EnvKeyProvider, FileKeyProvider, KeyProvider, ProviderError};
 // #[cfg(any(test, feature = "test-util"))]
 // pub use key_provider::StaticKeyProvider;
 // pub use layer::{ ... };
-// pub use store::{ ... };
-// #[cfg(any(test, feature = "credential-in-memory"))]
-// pub use memory::InMemoryStore;
